+++ changelog.md
@@ -3,6 +3,10 @@
 - Add coal coke to the "minecraft:coals" tag, allowing it to be used in crafting campfires (BluSunrize)
     - Unfortunately, vanilla torches don't use that tag, so they can't be made with coke. It's very silly.
 - Allow the hammer, wirecutters and faraday suit to be repaired with their respective materials (BluSunrize)
+- Add new tags to interact with arc furnace recycling:
+    - "immersiveengineering:recycling/whitelist" -> Items to be made recycle-able
+    - "immersiveengineering:recycling/blacklist" -> Items to exclude from recycling
+    - "immersiveengineering:recycling/ignored_components" -> Items to ignore as components (such as tools used in crafting)
 - Restructure some manual entries for better flow and no weird page breaks (BluSunrize)
 - Fix display of fluid tanks in JEI (BluSunrize)
 - Fix container items being lost in the alloy kiln (Malte)